+++ src/service/native-plugin-list.ts
@@ -1,5 +1,4 @@
 export interface QueryPluginInfo {
-  directory?: string;
   name?: string;
   version?: string;
   valid: boolean;